Move the SNI resolver, cert map and reload logic into huginn-certs
