+++ huginn-proxy-lib/src/tls/mod.rs
@@ -1,14 +1,14 @@
 pub mod acceptor;
-pub mod cert_resolver;
+pub mod cert_reload;
 pub mod cipher_suites;
 pub mod curves;
 pub mod metrics;
 pub mod session_resumption;
 pub mod setup;
 pub use acceptor::build_server_config_with_resolver;
-pub use cert_resolver::{CertReloadReport, DynamicCertResolver};
+pub use cert_reload::{cert_entries_from_domains, reload_certs};
 pub use cipher_suites::{is_cipher_suite_supported, supported_cipher_suites};
 pub use curves::{is_curve_supported, supported_curves};
-pub use huginn_certs::{cert_chain_hash, ServerCertsKeys};
+pub use huginn_certs::{cert_chain_hash, CertReloadReport, DynamicCertResolver, ServerCertsKeys};
 pub use metrics::{extract_tls_info, record_tls_handshake_metrics};
 pub use setup::build_tls_acceptor;
