+++ huginn-proxy-lib/tests/helpers.rs
@@ -1,4 +1,3 @@
-use std::fs;
 use std::path::PathBuf;
 use std::sync::OnceLock;
 use std::time::{Duration, SystemTime, UNIX_EPOCH};
@@ -12,28 +11,6 @@ pub fn tmp_path(name: &str) -> PathBuf {
         .unwrap_or_else(|_| Duration::from_secs(0))
         .as_nanos();
     std::env::temp_dir().join(format!("huginn-test-{nanos}-{name}"))
-}
-
-/// Generate valid test certificates using rcgen
-/// Returns paths to PEM files containing valid self-signed certificates
-/// Use for tests that need cryptographically valid certificates
-pub fn create_valid_test_cert(
-) -> Result<(PathBuf, PathBuf), Box<dyn std::error::Error + Send + Sync>> {
-    ensure_crypto_provider();
-    let cert_path = tmp_path("test.crt");
-    let key_path = tmp_path("test.key");
-
-    let subject_alt_names = vec!["localhost".to_string()];
-    let rcgen::CertifiedKey { cert, signing_key } =
-        rcgen::generate_simple_self_signed(subject_alt_names)?;
-
-    // Write certificate in PEM format
-    fs::write(&cert_path, cert.pem())?;
-
-    // Write private key in PEM format
-    fs::write(&key_path, signing_key.serialize_pem())?;
-
-    Ok((cert_path, key_path))
 }
 
 /// Generate valid test certificates as DER (for direct use with rustls)
